rag-chatbot/requirements.txt
# Core Dependencies
openai>=1.12.0
python-dotenv>=1.0.0

# LangChain for RAG
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-community>=0.0.20
langchain-text-splitters>=0.0.1

# Vector Store
chromadb>=0.4.22
sentence-transformers>=2.3.1

# Session Management
redis>=5.0.0

# Evaluation (Azure AI Evaluation SDK)
azure-ai-evaluation>=1.0.0
azure-identity>=1.15.0
promptflow>=1.8.0

# Guardrails
guardrails-ai>=0.4.0

# Tracing and Observability
opentelemetry-api>=1.22.0
opentelemetry-sdk>=1.22.0
opentelemetry-exporter-otlp>=1.22.0
opentelemetry-instrumentation-openai>=0.22.0

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Document Processing
pypdf>=4.0.0
python-docx>=1.1.0
beautifulsoup4>=4.12.0
markdown>=3.5.0

# API and Web
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0

# Utilities
tqdm>=4.66.0
tenacity>=8.2.0
